# openspec/specs/provider-switching/contracts/api.openapi.yaml
...
openapi: 3.1.0
info:
  title: Coffer — Provider Switching API
  version: 0.1.0
  description: The Provider Switching routes of Coffer's management API, generated from the daemon's Pydantic models. The behaviour behind them is this capability's spec.md; authentication, the error envelope and tracing are described once in the docs-site REST API reference. Every route except `GET /api/v1/daemon/status` requires the `X-Coffer-Token` header.
servers:
- url: http://127.0.0.1:8000
paths:
...
  /api/v1/providers:
    get:
      description: List all provider profiles.
      operationId: list_providers_api_v1_providers_get
      parameters:
      - in: header
        name: x-coffer-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Coffer-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderListOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The request failed validation (`CONFIG_INVALID`); the submitted values are not echoed back.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Any other error, as Coffer's error envelope.
      summary: List Providers
      tags:
      - providers
    post:
      description: Create a provider profile (422 when the secret source is invalid).
      operationId: create_provider_api_v1_providers_post
      parameters:
      - in: header
        name: x-coffer-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Coffer-Token
      - in: header
        name: x-coffer-actor
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Coffer-Actor
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The request failed validation (`CONFIG_INVALID`); the submitted values are not echoed back.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Any other error, as Coffer's error envelope.
      summary: Create Provider
      tags:
      - providers
  /api/v1/providers/detect-local:
    post:
      description: |-
        Which local model runtime answers where (spec provider-switching
        "Detect a local model runtime without changing it"). Read-only probes of
        loopback addresses only; nothing is pulled or loaded. A non-loopback URL
        is refused as 422.
      operationId: detect_local_api_v1_providers_detect_local_post
      parameters:
      - in: header
        name: x-coffer-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Coffer-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DetectLocalIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetectLocalOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The request failed validation (`CONFIG_INVALID`); the submitted values are not echoed back.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Any other error, as Coffer's error envelope.
      summary: Detect Local
      tags:
      - providers
  /api/v1/providers/use-builtin/{agent_type}:
    post:
      description: |-
        Switch every agent of this type back to its OWN built-in login: remove
        Coffer's projection from the native config and clear the active connection
        covering it. Idempotent — a no-op when the agent already runs built-in.
      operationId: use_builtin_provider_api_v1_providers_use_builtin__agent_type__post
      parameters:
      - in: path
        name: agent_type
        required: true
        schema:
          $ref: '#/components/schemas/AgentType'
      - in: header
        name: x-coffer-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Coffer-Token
      - in: header
        name: x-coffer-actor
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Coffer-Actor
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeactivateOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The request failed validation (`CONFIG_INVALID`); the submitted values are not echoed back.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Any other error, as Coffer's error envelope.
      summary: Use Builtin Provider
      tags:
      - providers
  /api/v1/providers/{uid}:
    delete:
      description: Delete a provider profile (404 if absent).
      operationId: delete_provider_api_v1_providers__uid__delete
      parameters:
      - in: path
        name: uid
        required: true
        schema:
          title: Uid
          type: string
      - in: header
        name: x-coffer-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Coffer-Token
      - in: header
        name: x-coffer-actor
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Coffer-Actor
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The request failed validation (`CONFIG_INVALID`); the submitted values are not echoed back.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Any other error, as Coffer's error envelope.
      summary: Delete Provider
      tags:
      - providers
    get:
      description: Get one provider profile (404 if absent).
      operationId: get_provider_api_v1_providers__uid__get
      parameters:
      - in: path
        name: uid
        required: true
        schema:
          title: Uid
          type: string
      - in: header
        name: x-coffer-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Coffer-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The request failed validation (`CONFIG_INVALID`); the submitted values are not echoed back.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Any other error, as Coffer's error envelope.
      summary: Get Provider
      tags:
      - providers
    patch:
      description: Partially update a provider profile.
      operationId: update_provider_api_v1_providers__uid__patch
      parameters:
      - in: path
        name: uid
        required: true
        schema:
          title: Uid
          type: string
      - in: header
        name: x-coffer-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Coffer-Token
      - in: header
        name: x-coffer-actor
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Coffer-Actor
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderPatch'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The request failed validation (`CONFIG_INVALID`); the submitted values are not echoed back.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Any other error, as Coffer's error envelope.
      summary: Update Provider
      tags:
      - providers
  /api/v1/providers/{uid}/activate:
    post:
      description: 'Switch: make this profile active for its wire format and project it.'
      operationId: activate_provider_api_v1_providers__uid__activate_post
      parameters:
      - in: path
        name: uid
        required: true
        schema:
          title: Uid
          type: string
      - in: header
        name: x-coffer-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Coffer-Token
      - in: header
        name: x-coffer-actor
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Coffer-Actor
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivateOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The request failed validation (`CONFIG_INVALID`); the submitted values are not echoed back.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Any other error, as Coffer's error envelope.
      summary: Activate Provider
      tags:
      - providers
  /api/v1/providers/{uid}/internal-default:
    post:
      description: |-
        Make this connection Coffer's internal-engine default (≤1 globally).

        Clears the flag on every other connection first, so setting a new default
        moves it off the previous one. 404 if the connection is absent.
      operationId: set_internal_default_provider_api_v1_providers__uid__internal_default_post
      parameters:
      - in: path
        name: uid
        required: true
        schema:
          title: Uid
          type: string
      - in: header
        name: x-coffer-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Coffer-Token
      - in: header
        name: x-coffer-actor
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Coffer-Actor
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The request failed validation (`CONFIG_INVALID`); the submitted values are not echoed back.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Any other error, as Coffer's error envelope.
      summary: Set Internal Default Provider
      tags:
      - providers
  /api/v1/providers/{uid}/transcribe-default:
    post:
      description: |-
        Make this connection the one Coffer transcribes speech on (≤1 globally).

        The twin of the route above, and deliberately a SECOND flag rather than a
        reuse of it: the two are different models, and a chat gateway commonly
        serves no ``/audio/transcriptions`` at all. Nothing falls back between
        them — with no connection marked here, Coffer transcribes nothing and hands
        the agent the audio file untouched. 404 if the connection is absent.
      operationId: set_transcribe_default_provider_api_v1_providers__uid__transcribe_default_post
      parameters:
      - in: path
        name: uid
        required: true
        schema:
          title: Uid
          type: string
      - in: header
        name: x-coffer-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Coffer-Token
      - in: header
        name: x-coffer-actor
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Coffer-Actor
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The request failed validation (`CONFIG_INVALID`); the submitted values are not echoed back.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Any other error, as Coffer's error envelope.
      summary: Set Transcribe Default Provider
      tags:
      - providers
...
components:
  schemas:
    ActivateOut:
      description: Result of a switch — which agents were written, which wire had none.
      properties:
        activated:
          title: Activated
          type: string
        projected:
          items:
            type: string
          title: Projected
          type: array
        protocol:
          $ref: '#/components/schemas/Protocol'
        skipped:
          items:
            type: string
          title: Skipped
          type: array
      required:
      - activated
      - protocol
      - projected
      - skipped
      title: ActivateOut
      type: object
...
    AgentType:
      description: Supported agent products.
      enum:
      - claude_code
      - codex
      title: AgentType
      type: string
...
    DeactivateOut:
      description: Result of switching an agent type back to its own built-in login.
      properties:
        agent_type:
          $ref: '#/components/schemas/AgentType'
        deprojected:
          items:
            type: string
          title: Deprojected
          type: array
        previous:
          anyOf:
          - type: string
          - type: 'null'
          title: Previous
      required:
      - agent_type
      - deprojected
      - previous
      title: DeactivateOut
      type: object
    DetectLocalIn:
      description: Probe one loopback URL, or — with none — each runtime's default port.
      properties:
        base_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Base Url
      title: DetectLocalIn
      type: object
    DetectLocalOut:
      properties:
        found:
          items:
            $ref: '#/components/schemas/LocalRuntimeOut'
          title: Found
          type: array
      required:
      - found
      title: DetectLocalOut
      type: object
...
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
      required:
      - error
      title: ErrorResponse
      type: object
...
    ProviderCreate:
      description: |-
        Create an LLM connection. For ``anthropic`` / ``openai`` / ``unknown``
        supply EXACTLY one of ``secret_value`` / ``secret_ref``; an ``ollama``
        connection has no key, so supply neither. WHICH agents the connection
        projects into is not set here: the new connection starts on the wire's own
        default scope and is re-targeted through the framework's scope surface
        (``PUT /api/v1/resources/{uid}/scope``), the same one every scoped
        kind uses. ``models`` curates which of the endpoint's models this connection
        offers downstream, each with its modality (``None`` ⇒ empty ⇒ no restriction).
      properties:
        base_url:
          minLength: 1
          title: Base Url
          type: string
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        local_runtime:
          anyOf:
          - $ref: '#/components/schemas/LocalRuntime'
          - type: 'null'
        models:
          anyOf:
          - items:
              $ref: '#/components/schemas/ProviderModel'
            type: array
          - type: 'null'
          title: Models
        name:
          maxLength: 64
          minLength: 1
          title: Name
          type: string
        protocol:
          $ref: '#/components/schemas/Protocol'
        secret_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Secret Ref
        secret_value:
          anyOf:
          - maxLength: 8192
            type: string
          - type: 'null'
          title: Secret Value
      required:
      - name
      - protocol
      - base_url
      title: ProviderCreate
      type: object
    ProviderListOut:
      properties:
        providers:
          items:
            $ref: '#/components/schemas/ProviderOut'
          title: Providers
          type: array
      required:
      - providers
      title: ProviderListOut
      type: object
...
    ProviderOut:
      description: |-
        An LLM connection as returned by the API (no secret).

        ``secret_ref`` is ``None`` for ``ollama`` connections (no key).
        ``compatible_agents`` is the CONFIGURED reach — the agent types this
        connection's per-agent scope (ADR per-agent-resource-scope) covers among the
        agents Coffer knows, not narrowed by ``enabled``, and empty for a keyless
        (ollama) connection — so the UI can filter agents without re-deriving it.
        It is READ-ONLY: it is reported here, and changed through the scope
        surface. ``models`` is the
        curated set of models this connection offers to every downstream picker, each
        carrying its modality; EMPTY means no restriction — the endpoint's whole
        catalogue. A picker takes the entries of the modality it serves, so a chat
        dropdown never offers an embedding or image model. ``internal_default``
        marks the connection Coffer's internal engine uses (at most one globally),
        ``transcribe_default`` the one it transcribes speech on — a separate flag
        because they are separate models and neither falls back to the other; and
        ``is_active`` marks the one currently projected. A connection may carry any
        combination.

        ``uid`` is the connection's identity and what every route here takes; the
        ``name`` beside it is the label, free to change through
        ``PATCH /api/v1/resources/{uid}`` without anything downstream noticing. That
        is the whole reason this kind no longer owns a rename operation of its own:
        nothing Coffer projects into an agent's config names the connection (the
        ``apiKeyHelper`` prints the agent's local proxy token), so a rename rewrites
        nothing (ADR resource-identity-is-an-immutable-uid).
      properties:
        base_url:
          title: Base Url
          type: string
        compatible_agents:
          items:
            $ref: '#/components/schemas/AgentType'
          title: Compatible Agents
          type: array
        created_at:
          format: date-time
          title: Created At
          type: string
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        enabled:
          title: Enabled
          type: boolean
        internal_default:
          title: Internal Default
          type: boolean
        is_active:
          title: Is Active
          type: boolean
        local_runtime:
          anyOf:
          - $ref: '#/components/schemas/LocalRuntime'
          - type: 'null'
        models:
          items:
            $ref: '#/components/schemas/ProviderModel'
          title: Models
          type: array
        name:
          title: Name
          type: string
        protocol:
          $ref: '#/components/schemas/Protocol'
        secret_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Secret Ref
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        transcribe_default:
          title: Transcribe Default
          type: boolean
        uid:
          title: Uid
          type: string
        updated_at:
          format: date-time
          title: Updated At
          type: string
      required:
      - uid
      - name
      - title
      - protocol
      - base_url
      - secret_ref
      - compatible_agents
      - models
      - is_active
      - internal_default
      - transcribe_default
      - enabled
      - description
      - created_at
      - local_runtime
      - updated_at
      title: ProviderOut
      type: object
    ProviderPatch:
      description: |-
        Partial update. ``secret_ref`` is immutable (it is the vault address
        the connection owns); ``protocol`` is not — the probe that guessed the wire
        can be wrong, so it is corrected in place rather than by re-entering the
        connection, key and all.

        The wire DOES decide one thing, though, so that correction is refused with
        409 ``PROVIDER_PROTOCOL_LOCKED_WHILE_ACTIVE`` while the connection is
        ``is_active``: an ``ollama`` connection covers no agent whatever its scope
        says. Moving the wire under a live projection would leave the native config
        already written with nothing that would ever take it off. Revert the agents to their
        built-in login, patch, then re-activate. An inactive connection patches
        freely.

        Re-targeting which agents the connection projects into is a scope edit, not
        a patch field (re-target then re-activate to re-project). ``models``
        replaces the curated set as a whole: ``None`` leaves it alone, ``[]`` clears
        the restriction.
      properties:
        base_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Base Url
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        models:
          anyOf:
          - items:
              $ref: '#/components/schemas/ProviderModel'
            type: array
          - type: 'null'
          title: Models
        protocol:
          anyOf:
          - $ref: '#/components/schemas/Protocol'
          - type: 'null'
        secret_value:
          anyOf:
          - maxLength: 8192
            type: string
          - type: 'null'
          title: Secret Value
      title: ProviderPatch
      type: object
...